requirements.txt
python-telegram-bot>=13.0
python-dotenv==0.19.2
aiofiles>=0.6.0